+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to `CloudwaysDeployManager` will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+
+
+### Added
+
+- Add config setting: `CW_DM_HTTP_TIMEOUT` to customize HTTP timeout to the Cloudways API
+
+
 ## [1.1.0] - 2024-10-16
 
 ### Added